extract _make_divisible to paddle.vision.models._utils
